+++ examples/server-backed/cmd/app/api_error.go
@@ -1,7 +1,0 @@
-package main
-
-type apiError string
-
-func (err apiError) Error() string {
-	return string(err)
-}